# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/carbon/footprint/calculate": {
            "post": {
                "description": "รับค่ากิจกรรมและระยะทาง แล้วคืนค่า Carbon Footprint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Carbon"
                ],
                "summary": "คำนวณ Carbon Footprint",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CarbonRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CarbonResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/carbon/footprint/calculate/weight": {
            "post": {
                "description": "รับค่ากิจกรรม, ระยะทาง และน้ำหนัก แล้วคืนค่า Carbon Footprint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Carbon"
                ],
                "summary": "คำนวณ Carbon Footprint (พาหนะที่มีน้ำหนัก)",
                "parameters": [
                    {
                        "description": "Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CarbonRequestWithWeight"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CarbonResponseWithWeight"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.CarbonRequest": {
            "type": "object",
            "properties": {
                "activity_type": {
                    "type": "string",
                    "example": "transportation"
                },
                "distance_km": {
                    "type": "number",
                    "example": 100
                },
                "fuel_type": {
                    "type": "string",
                    "example": "gasoline"
                },
                "vehicle_type": {
                    "type": "string",
                    "example": "car"
                }
            }
        },
        "domain.CarbonRequestWithWeight": {
            "type": "object",
            "properties": {
                "activity_type": {
                    "type": "string",
                    "example": "transportation"
                },
                "distance_km": {
                    "type": "number",
                    "example": 1000
                },
                "fuel_type": {
                    "type": "string",
                    "example": "jet_fuel"
                },
                "vehicle_type": {
                    "type": "string",
                    "example": "airplane"
                },
                "weight": {
                    "type": "number",
                    "example": 80000
                }
            }
        },
        "domain.CarbonResponse": {
            "type": "object",
            "properties": {
                "activity_type": {
                    "type": "string",
                    "example": "transportation"
                },
                "carbon_emission_kg": {
                    "type": "number",
                    "example": 25.4
                }
            }
        },
        "domain.CarbonResponseWithWeight": {
            "type": "object",
            "properties": {
                "activity_type": {
                    "type": "string",
                    "example": "transportation"
                },
                "carbon_emission_kg": {
                    "type": "number",
                    "example": 20000
                },
                "weight_tons": {
                    "type": "number",
                    "example": 80
                }
            }
        },
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        }
    }
}